basic functionality working!
Need to add timeouts (8min is so long, I don't remember that).

Need to add parsing of posts to self posts instead of links. This is because
when we have world+open class as the same show, it's annoying to post two
recaps links with an 8min delay. I think a better format overall is to post
it as a self-post with both groups there.

Need to add a better README and cronjob script.